Restructure /Code dir
