+++ packages/code/src/Code/Code.types.ts
@@ -1,0 +1,13 @@
+export const ScrollState = {
+  None: 'none',
+  Left: 'left',
+  Right: 'right',
+  Both: 'both',
+} as const;
+
+export type ScrollState = typeof ScrollState[keyof typeof ScrollState];
+
+export type DetailedElementProps<T> = React.DetailedHTMLProps<
+  React.HTMLAttributes<T>,
+  T
+>;
